fix: json LD normalizer will now not interfere with operation of other normalizers when useed through the symfony serializer
I changed the logic in JsonLdMapNormaliser to stop it trying to (de)normalise things if a different format was specified, because that was interfering with the deserialisation of plain JSON objects. I changed the JsonLdNormaliser as well for consistency, and because I could foresee that allowing it to continue trying to (de)normalise when the format was specified as somthing other than "json-ld" was going to be trouble.